requirements.txt
pytest==7.4.3
pytest-playwright==0.4.3
pytest-xdist==3.5.0
pytest-html==4.1.1
pytest-timeout==2.2.0
playwright==1.40.0
allure-pytest==2.13.2
python-dotenv==1.0.0
